+++ Scene/Title.cpp
@@ -393,7 +393,7 @@ void Title::Draw()const
 				}
 				if (swap_anim_timer <= 30)
 				{
-					DrawFormatStringF(menu_location[i].x + (menu_size[i].width / 2) - (GetDrawStringWidth(menu_string[i], strlen(menu_string[i])) / 2), menu_location[i].y + (menu_size[i].height / 2) - 24, 0xffffff, "%s", menu_string[i]);
+					DrawFormatStringF(menu_location[i].x + (menu_size[i].width / 2) - (GetDrawStringWidth(menu_string[i], (int)strlen(menu_string[i])) / 2), menu_location[i].y + (menu_size[i].height / 2) - 24, 0xffffff, "%s", menu_string[i]);
 					//選択中のメニューに交換カーソルを出す
 					if (current_menu == i)
 					{
